+++ packages/plugins/eslint-plugin-react-core/src/rules/avoid-shorthand-boolean.spec.ts
@@ -1,0 +1,34 @@
+import { allValid, ruleTester } from "../../../../../test";
+import rule, { RULE_NAME } from "./avoid-shorthand-boolean";
+
+ruleTester.run(RULE_NAME, rule, {
+  invalid: [
+    {
+      code: "<App foo />",
+      errors: [{
+        data: { propName: "foo" },
+        messageId: "AVOID_SHORTHAND_BOOLEAN",
+      }],
+    },
+    {
+      code: "<App foo bar />",
+      errors: [
+        {
+          data: { propName: "foo" },
+          messageId: "AVOID_SHORTHAND_BOOLEAN",
+        },
+        {
+          data: { propName: "bar" },
+          messageId: "AVOID_SHORTHAND_BOOLEAN",
+        },
+      ],
+    },
+  ],
+  valid: [
+    ...allValid,
+    "<App foo={true} />",
+    "<App foo={true} bar={true} />",
+    "<App foo={false} bar={false} />",
+    "<App foo={false} bar={false} baz={false} />",
+  ],
+});