+++ packages/plugins/eslint-plugin-react-core/src/rules/avoid-shorthand-boolean.ts
@@ -1,11 +1,10 @@
-import { NodeType } from "@eslint-react/ast";
 import { getPropName } from "@eslint-react/jsx";
 import type { ESLintUtils } from "@typescript-eslint/utils";
 import type { ConstantCase } from "string-ts";
 
 import { createRule } from "../utils";
 
-export const RULE_NAME = "prefer-shorthand-boolean";
+export const RULE_NAME = "avoid-shorthand-boolean";
 
 export type MessageID = ConstantCase<typeof RULE_NAME>;
 
@@ -14,11 +13,10 @@ export default createRule<[], MessageID>({
     type: "problem",
     docs: {
       description: "enforce boolean attributes notation in JSX",
-      recommended: "recommended",
       requiresTypeChecking: false,
     },
     messages: {
-      PREFER_SHORTHAND_BOOLEAN: "Prefer using shorthand boolean attribute '{{propName}}'",
+      AVOID_SHORTHAND_BOOLEAN: "Avoid using shorthand boolean attribute '{{propName}}'",
     },
     schema: [],
   },
@@ -28,17 +26,15 @@ export default createRule<[], MessageID>({
       JSXAttribute(node) {
         const { value } = node;
         const propName = getPropName(node);
-        const hasValueTrue = value?.type === NodeType.JSXExpressionContainer
-          && value.expression.type === NodeType.Literal
-          && value.expression.value === true;
-        if (!hasValueTrue) return;
-        context.report({
-          data: {
-            propName,
-          },
-          messageId: "PREFER_SHORTHAND_BOOLEAN",
-          node,
-        });
+        if (value === null) {
+          context.report({
+            data: {
+              propName,
+            },
+            messageId: "AVOID_SHORTHAND_BOOLEAN",
+            node,
+          });
+        }
       },
     };
   },