+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## v1.5.14 (Draft)
 
+### ✨ New
+
+- Add rule `react/avoid-shorthand-boolean`.
+- Add rule `react/avoid-shorthand-fragment`.
+
 ### 🐞 Fixes
 
 - Fix rule `react-dom/no-missing-button-type` false positive when using `type` attribute in a JSX expression.